Added feature to show new reservation count

+++ hsr_admin/views.py
@@ -26,6 +26,15 @@ from .forms import (
 # Create your views here.
 
 
+
+def new_reservations():
+    return Reservation.objects.filter(viewed = False).count()
+
+def context_parse(context):
+    context["status__"] = new_reservations()
+
+    return context
+
 def index(request: HttpRequest):
     if request.method.upper() == "POST":
         form = LoginForm(request.POST)
@@ -62,10 +71,12 @@ def index(request: HttpRequest):
 def home(request: HttpRequest):
     if not request.user.is_authenticated:
         return redirect(index)
-    return render(request, template_name="hsr_admin/dashboard.html", context={})
+    return render(request, template_name="hsr_admin/dashboard.html", context= context_parse({}))
 
 
 def category_list(request: HttpRequest):
+
+
     if not request.user.is_authenticated:
         return redirect(index)
 
@@ -73,15 +84,20 @@ def category_list(request: HttpRequest):
     return render(
         request,
         template_name="hsr_admin/categories.html",
-        context={"categories": _categories},
+        context=context_parse({"categories": _categories}),
     )
 
 
 def new_category(request: HttpRequest):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     if request.method == "GET":
         return render(
             request,
             "hsr_admin/new_category.html",
+            context=context_parse({})
         )
 
     form = CategoryForm(request.POST, request.FILES)
@@ -90,7 +106,7 @@ def new_category(request: HttpRequest):
         return render(
             request,
             "hsr_admin/new_category.html",
-            context={"message": "Please provide all the fields", "form": request.POST},
+            context=context_parse({"message": "Please provide all the fields", "form": request.POST}),
         )
     category = form.save()
 
@@ -109,13 +125,21 @@ def new_category(request: HttpRequest):
 
 
 def reservation_list(request: HttpRequest):
-    reservations = Reservation.objects.all()
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
+    reservations = Reservation.objects.all().order_by("-id")
     return render(
-        request, "hsr_admin/reservation_list.html", {"reservations": reservations}
+        request, "hsr_admin/reservation_list.html", context_parse({"reservations": reservations})
     )
 
 
 def room_list(request: HttpRequest):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     form = RoomFilterForm(request.GET)
 
     if form.is_valid():
@@ -147,16 +171,20 @@ def room_list(request: HttpRequest):
     return render(
         request,
         "hsr_admin/rooms.html",
-        {
+        context_parse({
             "categories": categories,
             "rooms": rooms,
             "colors": _colors,
             "form": request.GET,
-        },
+        }),
     )
 
 
 def view_room(request: HttpRequest, pk):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     try:
         room = Room.objects.get(pk=pk)
         categories = RoomCategory.objects.all()
@@ -167,12 +195,12 @@ def view_room(request: HttpRequest, pk):
                 return render(
                     request,
                     "hsr_admin/view_room.html",
-                    {
+                    context_parse({
                         "room": room,
                         "reservations": room.reservations.all(),
                         "categories": categories,
                         "error": "Please provide all field",
-                    },
+                    }),
                 )
 
             data = form.data
@@ -186,28 +214,32 @@ def view_room(request: HttpRequest, pk):
             return render(
                 request,
                 "hsr_admin/view_room.html",
-                {
+                context_parse({
                     "room": room,
                     "reservations": room.reservations.all(),
                     "categories": categories,
                     "success": "Room detail Updated",
-                },
+                }),
             )
 
         return render(
             request,
             "hsr_admin/view_room.html",
-            {
+            context_parse({
                 "room": room,
                 "reservations": room.reservations.all(),
                 "categories": categories,
-            },
+            }),
         )
     except Room.DoesNotExist:
         return redirect("hsr_admin:room_list")
 
 
 def view_category(request: HttpRequest, pk):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     try:
         category = RoomCategory.objects.get(pk=pk)
 
@@ -217,11 +249,11 @@ def view_category(request: HttpRequest, pk):
                 return render(
                     request,
                     "hsr_admin/view_category.html",
-                    {
+                    context_parse({
                         "rooms": category.rooms.all(),
                         "category": category,
                         "error": "Please provide all field",
-                    },
+                    }),
                 )
 
             data = form.cleaned_data
@@ -234,11 +266,11 @@ def view_category(request: HttpRequest, pk):
             return render(
                 request,
                 "hsr_admin/view_category.html",
-                {
+                context_parse({
                     "rooms": category.rooms.all(),
                     "category": category,
                     "success": "CAtegory detail Updated",
-                },
+                }),
             )
         elif request.method == "PUT":
 
@@ -247,46 +279,65 @@ def view_category(request: HttpRequest, pk):
         return render(
             request,
             "hsr_admin/view_category.html",
-            {
+            context_parse({
                 "rooms": category.rooms.all(),
                 "category": category,
-            },
+            }),
         )
     except Room.DoesNotExist:
         return redirect("hsr_admin:category_list")
 
 
 def delete_reservation(request, pk):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     reservation = Reservation.objects.get(pk=pk)
+
+    reservation.viewed = True
+    reservation.save()
+
 
     if request.method == "POST":
         reservation.delete()
         return redirect("hsr_admin:reservation_list")
 
     return render(
-        request, "hsr_admin/delete_reservation.html", {"reservation": reservation}
+        request, "hsr_admin/delete_reservation.html",context_parse( {"reservation": reservation})
     )
 
 
 def view_reservation(request, pk):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     reservation = get_object_or_404(Reservation, pk=pk)
 
+    reservation.viewed = True
+    reservation.save()
+
     return render(
-        request, "hsr_admin/view_reservation.html", {"reservation": reservation}
+        request, "hsr_admin/view_reservation.html",context_parse( {"reservation": reservation})
     )
 
 
 def check_availability(request: HttpRequest):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     form = CheckAvailabilityForm(request.POST)
 
     if not form.is_valid():
         return render(
             request,
             "hsr_admin/new_reservation.html",
-            context={
+            context=context_parse({
                 "message": "Please provide the reservation arrival and departure date and date",
                 "form": request.POST,
-            },
+            }),
         )
     data = form.cleaned_data
 
@@ -297,20 +348,20 @@ def check_availability(request: HttpRequest):
         return render(
             request,
             "hsr_admin/new_reservation.html",
-            context={
+            context=context_parse({
                 "message": "Please provide the reservation arrival and departure date and date",
                 "form": request.POST,
-            },
+            }),
         )
 
     if departure <= arrival:
         return render(
             request,
             "hsr_admin/new_reservation.html",
-            context={
+            context=context_parse({
                 "message": "Please departure date most be be ahead of arrival date",
                 "form": request.POST,
-            },
+            }),
         )
 
     rooms = RoomManager.available_rooms(None, arrival, departure)
@@ -332,17 +383,21 @@ def check_availability(request: HttpRequest):
     return render(
         request,
         "hsr_admin/new_reservation.html",
-        context={
+        context=context_parse({
             "adding": True,
             "form": request.POST,
             "categories": categories,
             "rooms": rooms,
             "colors": _colors,
-        },
+        }),
     )
 
 
 def new_reservation(request: HttpRequest):
+
+    if not request.user.is_authenticated:
+        return redirect(index)
+    
     if request.method == "POST":
         form = CheckAvailabilityForm(request.POST)
         customer_form = CustomerForm(request.POST)
@@ -352,10 +407,10 @@ def new_reservation(request: HttpRequest):
             return render(
                 request,
                 "hsr_admin/new_reservation.html",
-                context={
+                context=context_parse({
                     "message": "1 Please provide the reservation arrival and departure date and date",
                     "form": request.POST,
-                },
+                }),
             )
 
         # Check the reservation Form
@@ -368,20 +423,22 @@ def new_reservation(request: HttpRequest):
             return render(
                 request,
                 "hsr_admin/new_reservation.html",
-                context={
+                context=context_parse({
                     "message": "2 Please provide the reservation arrival and departure date and date",
                     "form": request.POST,
-                },
+                }),
             )
 
         if departure <= arrival:
             return render(
                 request,
                 "hsr_admin/new_reservation.html",
-                context={
+                context=context_parse(
+                {
                     "message": "3 Please departure date most be be ahead of arrival date",
                     "form": request.POST,
-                },
+                }
+                ),
             )
 
         if not customer_form.is_valid():
@@ -404,13 +461,13 @@ def new_reservation(request: HttpRequest):
             return render(
                 request,
                 "hsr_admin/new_reservation.html",
-                context={
+                context=context_parse({
                     "adding": True,
                     "form": request.POST,
                     "categories": categories,
                     "rooms": rooms,
                     "colors": _colors,
-                },
+                }),
             )
 
         if not booking_form.is_valid():
@@ -433,13 +490,13 @@ def new_reservation(request: HttpRequest):
             return render(
                 request,
                 "hsr_admin/new_reservation.html",
-                context={
+                context=context_parse({
                     "adding": True,
                     "form": request.POST,
                     "categories": categories,
                     "rooms": rooms,
                     "colors": _colors,
-                },
+                }),
             )
 
         customer_data = customer_form.cleaned_data
@@ -461,13 +518,13 @@ def new_reservation(request: HttpRequest):
             return render(
                 request,
                 "hsr_admin/new_reservation.html",
-                context={
+                context=context_parse({
                     "adding": True,
                     "form": request.POST,
                     "categories": categories,
                     "rooms": rooms,
                     "colors": _colors,
-                },
+                }),
             )
 
         reservation = Reservation(
@@ -489,4 +546,4 @@ def new_reservation(request: HttpRequest):
     else:
         form = ReservationForm()
 
-    return render(request, "hsr_admin/new_reservation.html", {"form": form})
+    return render(request, "hsr_admin/new_reservation.html",context_parse( {"form": form}))
